# backend/social-service/swagger.json
{"openapi":"3.0.0","paths":{"/api/friends/request":{"post":{"operationId":"FriendsController_sendFriendRequest","summary":"Send a friend request","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendFriendRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Friends"],"security":[{"bearer":[]}]}},"/api/friends/accept/{requestId}":{"post":{"operationId":"FriendsController_acceptFriendRequest","summary":"Accept a friend request","parameters":[{"name":"requestId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Friends"],"security":[{"bearer":[]}]}},"/api/friends/decline/{requestId}":{"post":{"operationId":"FriendsController_declineFriendRequest","summary":"Decline a friend request","parameters":[{"name":"requestId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Friends"],"security":[{"bearer":[]}]}},"/api/friends/{friendId}":{"delete":{"operationId":"FriendsController_removeFriend","summary":"Remove a friend","parameters":[{"name":"friendId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Friends"],"security":[{"bearer":[]}]}},"/api/friends":{"get":{"operationId":"FriendsController_getFriends","summary":"Get friends list","parameters":[{"name":"page","required":false,"in":"query","description":"Page number for pagination","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"default":20,"type":"number"}},{"name":"status","required":false,"in":"query","description":"Filter by friend status","schema":{"default":"all","enum":["online","offline","all"],"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Friends"],"security":[{"bearer":[]}]}},"/api/friends/requests":{"get":{"operationId":"FriendsController_getFriendRequests","summary":"Get pending friend requests","parameters":[],"responses":{"200":{"description":""}},"tags":["Friends"],"security":[{"bearer":[]}]}},"/api/friends/search":{"get":{"operationId":"FriendsController_searchUsers","summary":"Search for users by username","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Friends"],"security":[{"bearer":[]}]}},"/api/messages":{"post":{"operationId":"MessagesController_sendMessage","summary":"Send a message to a friend","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMessageDto"}}}},"responses":{"201":{"description":""}},"tags":["Messages"],"security":[{"bearer":[]}]}},"/api/messages/conversations":{"get":{"operationId":"MessagesController_getConversations","summary":"Get a list of conversations","parameters":[],"responses":{"200":{"description":""}},"tags":["Messages"],"security":[{"bearer":[]}]}},"/api/messages/conversations/{friendId}":{"get":{"operationId":"MessagesController_getConversation","summary":"Get conversation history with a friend","parameters":[{"name":"friendId","required":true,"in":"path","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","description":"Page number for pagination","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of messages per page","schema":{"default":50,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Messages"],"security":[{"bearer":[]}]}},"/api/messages/{messageId}/read":{"put":{"operationId":"MessagesController_markAsRead","summary":"Mark a message as read","parameters":[{"name":"messageId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Messages"],"security":[{"bearer":[]}]}},"/api/status/online":{"put":{"operationId":"StatusController_setOnlineStatus","summary":"Set the user's status to online","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetStatusDto"}}}},"responses":{"204":{"description":""}},"tags":["Online Status"],"security":[{"bearer":[]}]}},"/api/status/offline":{"put":{"operationId":"StatusController_setOfflineStatus","summary":"Set the user's status to offline","parameters":[],"responses":{"204":{"description":""}},"tags":["Online Status"],"security":[{"bearer":[]}]}},"/api/status/friends":{"get":{"operationId":"StatusController_getFriendsStatus","summary":"Get the online status of the user's friends","parameters":[],"responses":{"200":{"description":""}},"tags":["Online Status"],"security":[{"bearer":[]}]}},"/api/health":{"get":{"operationId":"HealthController_check","summary":"Check the health status of the service","parameters":[],"responses":{"200":{"description":""}},"tags":["Health"]}}},"info":{"title":"Social Service API","description":"API documentation for the Social Service of the Russian Game Platform","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SendFriendRequestDto":{"type":"object","properties":{"toUserId":{"type":"string","description":"The ID of the user to send the friend request to","example":"a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"},"message":{"type":"string","description":"Optional message to send with the friend request","example":"Hey, we played together yesterday!"}},"required":["toUserId"]},"SendMessageDto":{"type":"object","properties":{"toUserId":{"type":"string","description":"The ID of the user to send the message to","example":"a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"},"content":{"type":"string","description":"The content of the message","example":"Hello! How are you?","minLength":1,"maxLength":1000}},"required":["toUserId","content"]},"SetStatusDto":{"type":"object","properties":{"currentGame":{"type":"string","description":"The game the user is currently playing","example":"Cyberpunk 2077"}}}}}}